requirements.txt
backports.csv==1.0.5
certifi==2019.3.9
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.1.7
django-import-export==1.2.0
et-xmlfile==1.0.1
jdcal==1.4
mkl-fft==1.0.10
mkl-random==1.0.2
numpy==1.16.2
odfpy==1.4.0
openpyxl==2.6.1
pytz==2018.9
PyYAML==5.1
scikit-learn==0.20.3
scipy==1.2.1
tablib==0.13.0
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
